Add files via upload
Joystick 0 can control the other functions. A is intake and conveyor. RT is shooting. B is raise the climb arm. Y is pull in the rope. LT reverses the conveyor and the climb arm when held together with their respective buttons. LB reverses the rope when held with the respective button. Joystick 1 can drive and reverse heading. The front of the robot is now the shooting side.